Project/requirements.txt
numpy ==1.13.3
pytorch==1.6.0
tqdm ==4.55.0
python==3.6.3